Fix for #39
The `Position` must be defined on the various `ast.XXX` elements otherwise when the schema
validation fails, a nil pointer error is raised because most of the gqlparser error
construction function expects the `Position` field to be set.

To overcome the problem, this PR uses a single shared & cached position object for all
`Position` fields. A second approach will use a single position object for the active
parse action.