Proposition to change the generated document variable's name suffix to `Op` instead of `Query`
The `Query` suffix is not ideal in my opinion since:
- It's weird to use `Query` suffix when the actual document type is a Mutation, and gives weird usage like `updateUser(gql.UpdateUserQuery)`.
- A Query naming suffix in the generate config is something that could prove to be wanted, lead to confusion with the "document"

A such, I propose to change the suffix from `Query` to `Op`. I thought about using `Document`
or `Doc` but `Op` felt more aligned with the fact that we range over a list of operations.

This can be seen as a backward breaking change since upon a re-run of the gqlgenc binary,
some variables will not have the same name.

Alternatives:
- Keep `Query` as the default suffix but allow a rename through the `GenerateConfig` object.
- Perform the necessary change but generate also and alias ending in `Query`.
- Switch to `Op` as the default suffix but allow a rename through the `GenerateConfig` object to ease transition.
- Switch to `Op` directly without further changes.

Let me know what you think.