[CARESERV-928] Switch to fork of gqlgen to fix type generation conflict
[CARESERV-928](https://grandrounds.atlassian.net/browse/CARESERV-928)

I don't actually know if I need to do this in this repo. bobs uses both
gqlgenc and gqlgen. We're attempting to fork gqlgen, so I'm adding that
override to gqlgenc too. I don't know if go packages are smart enough to
resolve the override for both gqlgenc and bobs, which is why I'm adding
it here too