Keep `cfg.SchemaFilename` nil if "files" is empty
Add a condition to only set `cfg.SchemaFilename` if "files" has elements. In the LoadSchema function, whether loadLocalSchema or loadRemoteSchema is executed depends on whether `cfg.SchemaFilename` is nil or not, so if you want to read schema from a remote server, use `cfg SchemaFilename` is expected to be nil.